agent/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langgraph>=1.1.0
langchain-core>=0.3.0
huggingface_hub>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.7.0
